avanza-portal/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
requests==2.31.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
httpx==0.27.0
resend==0.7.2
slowapi==0.1.9
email-validator==2.2.0